refactor(parser): split lexer_state and errors into parser_pkg/ submodules
Phase 1B Batch 9 (Moves 21-22 from docs/PHASE1-PLAN.md, the final
Phase 1B batch).  Extract the token-cursor helpers and the
error-construction helpers into two new mixin modules.

* lexer_state.py — _at_end, _current, _advance, _match,
  _peek_ahead, _skip_newlines, _has_blank_line,
  _bracket_contains_slash.  Every rule method in parser_pkg
  reaches the token stream through these helpers.
* errors.py — _reject_stray_slash, _reject_reserved_decl_name.
  ParserError itself remains in parser_pkg/_base.py because the
  other mixins consume it from there.

Gate 9 verified:
* make check — clean.
* make test-e2e — 159/159 byte-identical.
* make refactor-diff — 190/190 byte-identical vs main.

parser.py shrinks from 948 → 846 lines (cumulative 86.7%).
Phase 1B is complete.
