+++ src/txt2tex/parser_pkg/__init__.py
@@ -4,7 +4,9 @@ from __future__ import annotations
 
 from txt2tex.parser_pkg import (
     algebra,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    errors,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     expressions,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    lexer_state,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     paragraphs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     proofs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     schemas,  # noqa: F401  # pyright: ignore[reportUnusedImport]
